backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.2
pydantic-settings==2.13.1
python-dotenv==1.2.2
supabase==2.28.3
PyJWT==2.12.1
httpx==0.28.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.84
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
openai==1.109.1
requests==2.33.1
python-multipart==0.0.20
